figured out import problem
python imports finally figured out. this was incredibly not what i
wanted my life to work. the root of the package has to be the same
name as the package folders themselves which was not clear to me.
so the fix is, i created a new directory which is this current one
and moved all the package code into this one. and tried the testing
and it worked just fine, so this is where the devel process is going to
be now. i should have done this a long time ago but i just did it.

the shoot directory is going to contain only code that relates to the
distributed chat system i wanted to create in the first place.

next steps, i'm merging this branch into master and do all devel work
from the master branch, i do not want uneccessary headache from
dealing  with git branch and merge problems.