complete dialing and listening functions
The functions for dialing and listening for sockets have been completed.
What is left is testing and refactoring to see if it works as expected.
the code looks somehow neat i think, except for the occassional oh sh!t
i totally forget to add this here and i have to add it now and
everything becomes clumsy.